Add support for running the integration suite from riak_test.
The purpose of this custom suite is not to exercise the client, but to
throw client requests at Riak and see if they fail because of changes
in the server-side code.

A "dummy" test server is employed that uses a node already created by
an external process like riak_test. Turning this on relies on setting
the appropriate environment variables before running `rake riak_test`.
It assumes that the given node will:

1) Be at least Riak 1.2.0 or later.
2) Use the memory backend, with the `test` flag enabled.
3) Enable search and use the search backend included in
   riak-client/erl_src (although this is not strictly necessary in the
   suite's current state).
4) Use the default `PIPE_DIR` of `/tmp/$RUNNER_BASE_DIR/`.