Guard against an empty value when loading content from ProtoBufs
Under the HTTP client backend, the raw data for an object was
only set when present in the response. On the ProtoBuf backend,
the raw data is always set regardless - even when there is no
data or content type. This behavior results in a NoMethodError
exception when deserialzing the raw data on an `RContent` instance.

By only setting the raw data property when a value is present,
the deserialization step will be skipped and an exception won't
be raised.

Ref: https://github.com/basho/riak-ruby-client/blob/v1.4.4.1/lib/riak/client/http_backend/object_methods.rb#L90-L109